test(core): migrate UpdateLifecycleTests (Wave C)
27 tests, 61 failure sites preserved 1:1. The six commands[0] case-pattern guards convert to Issue.record + return.
